Data editor

+++ src/locaria/components/admin/redux/slices/adminPagesSlice.js
@@ -14,9 +14,13 @@ export const adminPagesSlice = createSlice({
 		feature: undefined,
 		editor: undefined,
 		awaitUI: false,
-		user:false
+		user:false,
+		token: undefined
 	},
 	reducers: {
+		setToken:(state,action) => {
+			state.token=action.payload;
+		},
 		clearUI:(state,actions) => {
 			state.awaitUI=false;
 		},
@@ -39,8 +43,12 @@ export const adminPagesSlice = createSlice({
 			state.style = actions.payload;
 		},
 		setOverview: (state, actions) => {
-			actions.payload.total_updates=actions.payload['add_item']+actions.payload['delete_item']+actions.payload['update_item'];
-			state.overview = actions.payload;
+			if(actions.payload===undefined) {
+				state.overview = undefined;
+			} else {
+				actions.payload.total_updates = actions.payload['add_item'] + actions.payload['delete_item'] + actions.payload['update_item'];
+				state.overview = actions.payload;
+			}
 		},
 		updateStyle: (state,actions) => {
 			if(actions.payload.data)
@@ -87,6 +95,6 @@ export const adminPagesSlice = createSlice({
 })
 
 // Action creators are generated for each case reducer function
-export const {setPages, addPage, setPage, setStyle,updateStyle,setOverview,setFeature,clearUI,setEditor,setUser} = adminPagesSlice.actions
+export const {setPages, addPage, setPage, setStyle,updateStyle,setOverview,setFeature,clearUI,setEditor,setUser,setToken} = adminPagesSlice.actions
 
 export default adminPagesSlice.reducer
